Split grab-bag E2E specs
Break the secrets and workspace browser flows into surface-focused specs and move listener workflow setup into suite infrastructure. This keeps the same coverage while making each spec file describe one behavior area.

Fixes ENG-816
