+++ e2e/screens/workspaces/src/index.ts
@@ -3,6 +3,82 @@ import type {Page} from '@shipfox/playwright';
 
 type Locator = ReturnType<Page['locator']>;
 type FixtureUse<T> = (fixture: T) => Promise<void>;
+const LAST_WORKSPACE_KEY = 'shipfox.lastWorkspaceId';
+
+export class WorkspaceOnboardingScreen {
+  constructor(private readonly page: Page) {}
+
+  async gotoRoot(): Promise<void> {
+    await this.page.goto('/');
+  }
+
+  async gotoWorkspace(workspaceId: string): Promise<void> {
+    await this.page.goto(`/workspaces/${workspaceId}`);
+  }
+
+  heading(): Locator {
+    return this.page.getByRole('heading', {name: 'Create your workspace'});
+  }
+
+  workspaceNameField(): Locator {
+    return this.page.getByLabel('Workspace name');
+  }
+
+  createWorkspaceButton(): Locator {
+    return this.page.getByRole('button', {name: 'Create workspace'});
+  }
+
+  async createWorkspace(name: string): Promise<void> {
+    await this.workspaceNameField().fill(name);
+    await this.createWorkspaceButton().click();
+  }
+}
+
+export class WorkspaceHomeScreen {
+  constructor(private readonly page: Page) {}
+
+  async gotoRoot(): Promise<void> {
+    await this.page.goto('/');
+  }
+
+  async goto(workspaceId: string): Promise<void> {
+    await this.page.goto(`/workspaces/${workspaceId}`);
+  }
+
+  async gotoIntegrations(workspaceId: string): Promise<void> {
+    await this.page.goto(`/workspaces/${workspaceId}/integrations`);
+  }
+
+  async gotoSettings(workspaceId: string): Promise<void> {
+    await this.page.goto(`/workspaces/${workspaceId}/settings`);
+  }
+
+  settingsTab(): Locator {
+    return this.page.getByRole('tab', {name: 'Settings'});
+  }
+
+  currentWorkspaceId(): string | undefined {
+    return new URL(this.page.url()).pathname.split('/')[2];
+  }
+
+  async readMaybeLastWorkspaceId(): Promise<string | undefined> {
+    const raw = await this.page.evaluate(
+      (key) => window.localStorage.getItem(key),
+      LAST_WORKSPACE_KEY,
+    );
+    if (!raw) return undefined;
+    const parsed: unknown = JSON.parse(raw);
+    return typeof parsed === 'string' ? parsed : undefined;
+  }
+
+  async readLastWorkspaceId(): Promise<string> {
+    const workspaceId = await this.readMaybeLastWorkspaceId();
+    if (workspaceId === undefined) {
+      throw new Error(`localStorage[${LAST_WORKSPACE_KEY}] is not set`);
+    }
+    return workspaceId;
+  }
+}
 
 export class MembersSettingsScreen {
   private readonly shell: SettingsShell;
@@ -13,6 +89,10 @@ export class MembersSettingsScreen {
 
   async goto(workspaceId: string): Promise<void> {
     await this.shell.goto(workspaceId, 'members');
+  }
+
+  async gotoDefault(workspaceId: string): Promise<void> {
+    await this.page.goto(`/workspaces/${workspaceId}/settings`);
   }
 
   heading(): Locator {
@@ -99,6 +179,8 @@ export class InvitationAcceptScreen {
 export interface WorkspacesScreenFixtures {
   invitationAccept: InvitationAcceptScreen;
   membersSettings: MembersSettingsScreen;
+  workspaceHome: WorkspaceHomeScreen;
+  workspaceOnboarding: WorkspaceOnboardingScreen;
 }
 
 export const workspacesScreens = {
@@ -107,5 +189,11 @@ export const workspacesScreens = {
   },
   membersSettings: async ({page}: {page: Page}, use: FixtureUse<MembersSettingsScreen>) => {
     await use(new MembersSettingsScreen(page));
+  },
+  workspaceHome: async ({page}: {page: Page}, use: FixtureUse<WorkspaceHomeScreen>) => {
+    await use(new WorkspaceHomeScreen(page));
+  },
+  workspaceOnboarding: async ({page}: {page: Page}, use: FixtureUse<WorkspaceOnboardingScreen>) => {
+    await use(new WorkspaceOnboardingScreen(page));
   },
 };
